Convert to dotnet core
