+++ .gitignore
@@ -1,5 +1,5 @@
 *.mp3
 *.cache
 .vs
-obj/*
-
+obj/
+bin/
