--- odysseyDownloader.sln
@@ -1,23 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "odysseyDownloader", "odysseyDownloader\odysseyDownloader.csproj", "{9D2A06AB-3F57-4EB8-A27B-30C2000E6BC3}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x86 = Debug|x86
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9D2A06AB-3F57-4EB8-A27B-30C2000E6BC3}.Debug|x86.ActiveCfg = Debug|x86
-		{9D2A06AB-3F57-4EB8-A27B-30C2000E6BC3}.Debug|x86.Build.0 = Debug|x86
-		{9D2A06AB-3F57-4EB8-A27B-30C2000E6BC3}.Release|x86.ActiveCfg = Release|x86
-		{9D2A06AB-3F57-4EB8-A27B-30C2000E6BC3}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(MonoDevelopProperties) = preSolution
-		StartupItem = odysseyDownloader\odysseyDownloader.csproj
-	EndGlobalSection
-EndGlobal
