--- odysseyDownloader.userprefs
@@ -1,44 +1,0 @@
-﻿<Properties>
-  <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug|x86" ctype="Workspace" />
-  <MonoDevelop.Ide.Workbench ActiveDocument="odysseyDownloader/getFileInfo.cs" ctype="Workbench">
-    <Files>
-      <File FileName="odysseyDownloader/Main.cs" Line="22" Column="39" />
-      <File FileName="odysseyDownloader/getFileInfo.cs" Line="47" Column="1" />
-      <File FileName="odysseyDownloader/CGWebClient.cs" Line="53" Column="1" />
-      <File FileName="odysseyDownloader/multiDownload.cs" Line="18" Column="30" />
-      <File FileName="odysseyDownloader/processFile.cs" Line="91" Column="30" />
-    </Files>
-  </MonoDevelop.Ide.Workbench>
-  <MonoDevelop.Ide.DebuggingService.Breakpoints>
-    <BreakpointStore />
-  </MonoDevelop.Ide.DebuggingService.Breakpoints>
-  <MonoDevelop.Ide.DebuggingService.PinnedWatches ctype="PinnedWatchStore" />
-  <MonoDevelop.Ide.ItemProperties.odysseyDownloader ctype="Properties">
-    <MonoDevelop.Ide.CustomExecutionModes ctype="CustomExecutionModes">
-      <Data Name="Mono Soft Debugger (Custom)" Id="d32d2d87-89c2-43e0-ac64-8d7bd65e9b61" ModeId="Mono Soft Debugger">
-        <CommandData>
-          <Item Key="MonoDevelop.Ide.Execution.CustomArgsCustomizer">
-            <Value Arguments="-c 3 -config &quot;../config.xml&quot;" ctype="CustomArgsExecutionModeData">
-              <EnvironmentVariables />
-            </Value>
-          </Item>
-          <Item Key="MonoDevelop.Ide.Execution.MonoExecutionCustomizer">
-            <Value ctype="MonoExecutionParameters" />
-          </Item>
-        </CommandData>
-      </Data>
-      <Data Name="Odyssey Folder --all" Id="43407a80-ce82-49a0-b4f7-c902932eb118" ModeId="Mono Soft Debugger">
-        <CommandData>
-          <Item Key="MonoDevelop.Ide.Execution.CustomArgsCustomizer">
-            <Value Arguments="--all" WorkingDirectory="/home/steven/Music/Odyssey" ctype="CustomArgsExecutionModeData">
-              <EnvironmentVariables />
-            </Value>
-          </Item>
-          <Item Key="MonoDevelop.Ide.Execution.MonoExecutionCustomizer">
-            <Value ctype="MonoExecutionParameters" />
-          </Item>
-        </CommandData>
-      </Data>
-    </MonoDevelop.Ide.CustomExecutionModes>
-  </MonoDevelop.Ide.ItemProperties.odysseyDownloader>
-</Properties>
